Update React-related dependencies
The updated template `dependencies` match what currently gets pulled in
with the latest versions of create-react-app.

Some of the eslint `devDependencies` were included in (and conflicting
with) the version of react-scripts that my setup pulled in, so those
have been removed in preference of the versions transitively included in
`dependencies`.